// src/types/Responses.d.ts
export type PartyGetByIdResponse = {
  __v?: number;
  _id: string;
  avatarUrl?: string;
  /** WarEra user IDs of the party's council members. */
  councilMembers?: Array<string>;
  country?: string;
  createdAt?: string;
  description?: string;
  ethics?: Record<string, unknown>;
  /** WarEra user ID of the party leader/president. */
  leader?: string;
  /** WarEra user IDs of all party members (may include leader/treasurer/council). */
  members?: Array<string>;
  name?: string;
  primaryWinner?: string;
  region?: string;
  /** WarEra user ID of the party treasurer. */
  treasurer?: string;
  updatedAt?: string;
};